requirements.txt
greenlet==1.0.0
paho-mqtt==1.5.1
SQLAlchemy==1.4.1
